+++ src/helpers/index.ts
@@ -1,4 +1,3 @@
-export * from "./isEmpty";
 export * from "./createAuthToken";
 export * from "./createNavLinkStyler";
 export * from "./debounce";
@@ -7,6 +6,7 @@ export * from "./getPossessiveForm";
 export * from "./evm";
 export * from "./getTxUrl";
 export * from "./idParamToNum";
+export * from "./isEmpty";
 export * from "./logger";
 export * from "./maskAddress";
 export * from "./theme";
