+++ src/pages/Admin/Charity/Banking/RecipientDetails/types.ts
@@ -1,4 +1,4 @@
-import { CreateRecipientRequest } from "../types";
+import { CreateRecipientRequest } from "types/aws";
 import { OptionType } from "components/Selector";
 
 export type FormValues = Omit<CreateRecipientRequest, "details"> & {
