+++ src/pages/Admin/Charity/Banking/RecipientDetails/useRecipientForm/getDefaultValues.ts
@@ -1,5 +1,5 @@
-import { AccountRequirements } from "../../types";
 import { FormValues } from "../types";
+import { AccountRequirements } from "types/aws";
 import { undot } from "../dot";
 
 export default function getDefaultValues(
