+++ src/pages/Admin/Charity/Banking/RecipientDetails/useRecipientForm/useRecipientForm.ts
@@ -1,7 +1,7 @@
 import { yupResolver } from "@hookform/resolvers/yup";
 import { UseFormReturn, useForm } from "react-hook-form";
-import { AccountRequirements } from "../../types";
 import { FormValues } from "../types";
+import { AccountRequirements } from "types/aws";
 import createSchema from "./createSchema";
 import getDefaultValues from "./getDefaultValues";
