+++ src/services/aws/aws.ts
@@ -2,7 +2,6 @@ import { createApi, fetchBaseQuery, retry } from "@reduxjs/toolkit/query/react";
 import {
   ProfileUpdatePayload,
   VersionSpecificWalletProfile,
-  WiseRequest,
   isDeleteMsg,
 } from "../types";
 import {
@@ -136,17 +135,6 @@ export const aws = createApi({
         };
       },
     }),
-    /**
-     * The endpoint returns different types of results depending on the requests sent, but since it's impossible
-     * to specify a generic return argument here, we return `unknown` and cast the return value manually
-     */
-    wise: builder.mutation<unknown, WiseRequest>({
-      query: (payload) => ({
-        url: `/${v(1)}/wise`,
-        method: "POST",
-        body: payload,
-      }),
-    }),
   }),
 });
 
@@ -158,7 +146,6 @@ export const {
   useProfileQuery,
   useProgramQuery,
   useEditProfileMutation,
-  useWiseMutation,
 
   endpoints: {
     endowmentCards: { useLazyQuery: useLazyEndowmentCardsQuery },
