+++ src/services/aws/bankDetails/index.ts
@@ -1,0 +1,1 @@
+export * from "./useWiseMutationProxy";
