+++ src/types/aws/index.ts
@@ -1,5 +1,6 @@
 export * from "./common";
 export * from "./ap";
 export * from "./ap/registration";
+export * from "./ap/bankDetails";
 export * from "./apes";
 export * from "./apes/donation";
